large refactor
This does a bunch of aggregation so the same logic is shared across
multiple files, and the implementation of each of the key functions is
improved to make it much clearer what's going on.

Should be fewer edge cases related to filenames and tags etc. as well as
reusing some very similar functionality and building upon it, rather
than duplicating and slightly tweaking a function.